Add moving of components into sub folder
