+++ src/pages/VaccinePage.js
@@ -23,7 +23,7 @@ import { generalStrings, buttonStrings } from '../localization';
 import { APP_FONT_FAMILY, DARKER_GREY, BLACK } from '../globalStyles';
 import { gotoFridgeDetailPage, gotoNewSensorPage } from '../navigation/actions';
 import { AfterInteractions } from '../widgets/AfterInteractions';
-import { SensorHeader } from '../widgets/SensorHeader';
+import { SensorHeader } from '../widgets/SensorHeader/SensorHeader';
 import { selectSensors } from '../selectors/Entities/sensor';
 import temperature from '../utilities/temperature';
 import { BreachManUnhappy } from '../widgets/BreachManUnhappy';
