+++ src/widgets/SensorHeader/LastSensorDownload.js
@@ -5,17 +5,17 @@ import { StyleSheet } from 'react-native';
 import * as Animatable from 'react-native-animatable';
 import moment from 'moment';
 
-import { useIntervalReRender } from '../hooks';
-import { TextWithIcon } from './Typography/index';
-import { MILLISECONDS } from '../utilities/index';
-import { WifiIcon, WifiOffIcon } from './icons';
-import { MISTY_CHARCOAL } from '../globalStyles/index';
+import { useIntervalReRender } from '../../hooks';
+import { TextWithIcon } from '../Typography/index';
+import { MILLISECONDS } from '../../utilities/index';
+import { WifiIcon, WifiOffIcon } from '../icons';
+import { MISTY_CHARCOAL } from '../../globalStyles/index';
 import {
   selectIsDownloading,
   selectLastDownloadFailed,
   selectLastDownloadTime,
-} from '../selectors/Bluetooth/sensorDownload';
-import { generalStrings } from '../localization';
+} from '../../selectors/Bluetooth/sensorDownload';
+import { generalStrings } from '../../localization';
 
 const formatLastSyncDate = date => (date ? moment(date).fromNow() : generalStrings.not_available);
