+++ src/widgets/SensorHeader/SensorHeader.js
@@ -3,22 +3,22 @@ import PropTypes from 'prop-types';
 
 import { connect } from 'react-redux';
 import { StyleSheet } from 'react-native';
-import { TextWithIcon } from './Typography/index';
-import { BatteryIcon, CogIcon } from './icons';
+import { TextWithIcon } from '../Typography/index';
+import { BatteryIcon, CogIcon } from '../icons';
 import {
   APP_FONT_FAMILY,
   BLACK,
   DARKER_GREY,
   FINALISE_GREEN,
   MISTY_CHARCOAL,
-} from '../globalStyles/index';
-import { IconButton } from './IconButton';
-import { selectSensorState } from '../selectors/Entities/sensor';
-import { gotoEditSensorPage } from '../navigation/actions';
+} from '../../globalStyles/index';
+import { IconButton } from '../IconButton';
+import { selectSensorState } from '../../selectors/Entities/sensor';
+import { gotoEditSensorPage } from '../../navigation/actions';
 import { LastSensorDownload } from './LastSensorDownload';
 import { BlinkSensorButton } from './BlinkSensorButton';
-import { Circle } from './Circle';
-import { FlexRow } from './FlexRow';
+import { Circle } from '../Circle';
+import { FlexRow } from '../FlexRow';
 import { ExportTemperatureDataButton } from './ExportTemperatureDataButton';
 
 const formatBatteryLevel = batteryLevel => `${batteryLevel}%`;
